chore: split gateway CLI commands into separate files
